Add things-internals.ts that contains the reverse engineered Things database
